Move logic to separate file
